Replace fenced codeblocks with Kramdown syntax
This is based on the Kramdown docs
(http://kramdown.gettalong.org/syntax.html#fenced-code-blocks). I am not
a fan of this over the ```, but in order for code blocks to work the ~~~
syntax needs to be used. I will look into why this is and why it
deviates from GitHub Flavored Markdown.