Simplify the nav
This removes the Contact page from the nav and instead opts to have the
content live in the About page. This also removes the links to Twitter &
GitHub from the nav. They will live in the About page as well in their
own section.

The Projects page is now part of the nav, and that will be the page that
highlights the work I have done. The content there is not ready yet.

This makes the nav cleaner and more intentional.