Switch to Sass for styles
This moves all of the CSS into the _sass dir and changes their
extensions to .scss so they are interpreted as Sass files. This will
allow for better written styles and now compresses the output into 1
compressed stylesheet which is better for performance.

I used https://jekyllrb.com/docs/assets/ as the reference for using Sass
with Jekyll.