Add & configure jekyll-seo-tag
The jekyll-seo-tag plugin make it easy to specify metadata for search
engines and social networks throughout the site. This adds and sets the
config up for what it expects.

The page titles all changes because the plugin handles setting the title
and appends "- Brett Chalupa" after the title, so the pages do not need
to specify it as well.