Update license for code & content (MIT & CC)
This updates the license info for the code and content of the site to be
more liberal to allow people to share and use what I have made. I am
doing this so that my works are more accessible and potentially more
useful. I think both of the selected licenses will serve me well as I
continue to grow as a creator, and as this website grows.

The associated changes include:

- Adding a new License page that lays out the details
- Adding a link to the footer that links to the License page
- Adding a link to the About page about the license
- Updating the README to link to the License page