Display associated work collection items
This updates the work layout to loop through and output each document of
the associated collection, if there is one specified (via the
collection_label attr). It also adds the option to display an associated
collection title for the kind, e.g. Chapter, Story, Episode, etc.

This will be used for works like novels and short story collections to
link to each entry of that work.