Add initial chapter layout
This adds an inital chapter layout for rendering the sections of a work.
I decided to go with chapter since that is most analogous with a book,
which is what I mostly intend to publish. It'll still apply well to
other media with associated entries.

The chapter layout has the following features:

- Shows the content of the chapter
- Shows table of contents if there is more than one chapter
- Shows the next / prev chapter links
- Embeds Soundcloud if the id is present. I may make it lazy load in the
  future.

This will be used first for Brief Moments, and in the future it will be
used for other novels and works that have sections.

This commit also includes some styles to make it look nice.