Add additional info links to work layout
This makes it so that a work can specific additional info for enjoying
elsewhere, below the list of chapter and main content. To me those calls
of action are a bit lower in priority than enjoying the work, so this
should be nice. It also allows the links to be easily managed in the
front matter of the works.