Add Short Story layout & styles for prose
This adds a new layout for works which are short stories in prep for
adding Latency. This is more ideal for short stories that I think stand
on their own (versus BMOL which I don't want to show individually on the
Works page).

This also adds styles for prose, which justifies the content, indents
paragraphs, and styles the text to look more like a novel. I think this
will improve the reading experience.