Add newsletter & Subscribe page
This integrates the Gumroad newsletter sign up form in the site by
displaying it on the home page and newly added Subscribe page. The goal
is to make it easy for people to get updates on my works however they
want.

This is accomplished through using a newsletter snippet to display the
form where needed. The styles have also been tweaked to fit the theme of
the site.